Added command decorator and call action which both act on a registry global.
The  decorator serves to record functions and the name of their arguments to a registry.  recieves an action, json to act on and an argument source and parses the registry to get the needed information to complete the call.

Globals aren't so bad all the time, but I still don't feel great about this one.