Added command decorator, call, act_on_path and set_on_path.
The command decorator places functions and their argument list into a
global registery. The call function accepts a function name, a JSON
representation, and an object that has the arguments stored on it and
calls the stored function on the JSON with the appropriate arguments.
act_on_path calls a function at the end point of a JSON path.
set_on_path calls operator.setitem through act_on_path.